Update registerDebugSession event handler thread to take callback function.
- Relocate 'runAdaptorWith' to Adaptor.hs
- Add some comments for 'registerDebugSession'
